backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
marshmallow==3.21.3
webargs==8.4.0
gunicorn==22.0.0
redis==5.0.8
celery==5.3.6
langchain==0.2.16
langchain-openai==0.1.23
langchain-anthropic==0.1.23
openai==1.51.2
anthropic==0.34.2
pydub==0.25.1
librosa==0.10.2
soundfile==0.12.1
numpy>=1.26.0,<2.0.0
scipy>=1.11.0
requests==2.32.3
pytest==8.3.3
pytest-flask==1.3.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2
